Vista basica listado de casos, incorporación boton de interes, creación de aviso de interes

+++ database/migrations/2023_09_12_161128_create_notices_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('notices', function (Blueprint $table) {
             $table->id();
             $table->string('description');
-            $table->integer('contact_number');
+            $table->bigInteger('contact_number');
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
             $table->foreignId('post_id')->constrained()->onDelete('cascade');
             $table->timestamps();
